moved navigation bar classes to its correct directory
